Typescriptify more configuration editor stuff
